+++ CHANGES.md
@@ -1,6 +1,7 @@
 Changes coming in 0.7.4
 
 * Improved discoverability of other `java.jdbc` documentation [JDBC-160](https://dev.clojure.org/jira/browse/JDBC-160).
+* Optional specs updated with `:keywordize?` and `:connection-uri` changes from 0.7.2 and 0.7.3 releases.
 
 Changes in 0.7.3